+++ Examples/2.x/Example/Modules/Images/Sources/Images/ContentView.swift
@@ -29,6 +29,8 @@ public struct ContentView: View {
             // SwiftGen ussing:
             // Asset.Folder.duplicatedImage3.image
             // Asset.duplicatedImage1.image
+            // Asset.snakeCaseImage.image
+            Image("snake-case_image")
         }
         .padding()
     }